Refactor Sonarr and Radarr classes to share common functionality
